Remove copy resources functionality
The copying of resources is quite fragile and can actually be handled fine
in some limited cases as seen in the extra_data example. For more
advanced cases (i.e. where the referenced label != the dhall import path
epxression), this should be redone to match how imports are done for
freezing (i.e. environment variable with a conventional name to match
data label name)